con_x11: Make compatible with tiling window managers
Tiling window managers can force the window size to be larger than
supported, causing either an early abort at ConSetSize, or if compiled
with NDEBUG, a segmentation fault down the road.

Instead, just cap X and Y to their maximums. The only consequence is that
the window is not fully utilised.